# ChallengeAppApi/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ChallengeAppApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/challenges": {
      "get": {
        "tags": [
          "Challenges"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ChallengeEntity"
              }
            }
          }
        }
      }
    },
    "/challenges/{id}": {
      "get": {
        "tags": [
          "Challenges"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/challenges/create": {
      "post": {
        "tags": [
          "Challenges"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ChallengeModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/challenges/edit/{id}": {
      "put": {
        "tags": [
          "Challenges"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ChallengeModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/challenges/delete/{id}": {
      "delete": {
        "tags": [
          "Challenges"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "ChallengeEntity": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "minLength": 1,
          "type": "string"
        },
        "difficultyLevel": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "duration": {
          "format": "int32",
          "type": "integer"
        },
        "startedAt": {
          "format": "date-time",
          "type": "string"
        },
        "githubLink": {
          "type": "string"
        }
      }
    },
    "ChallengeModel": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "name": {
          "minLength": 1,
          "type": "string"
        },
        "difficultyLevel": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "duration": {
          "format": "int32",
          "type": "integer"
        },
        "startedAt": {
          "format": "date-time",
          "type": "string"
        },
        "githubLink": {
          "type": "string"
        }
      }
    }
  }
}